Fix warnings in aggregator.cc

+++ c/extras/aggregator.h
@@ -28,20 +28,20 @@ typedef std::unique_ptr<ex> ExPtr;
 
 class Aggregator {
   public:
-    Aggregator(int32_t, int32_t, int32_t, int32_t, int32_t, int32_t,
+    Aggregator(size_t, size_t, size_t, size_t, size_t, size_t,
                unsigned int, PyObject*, unsigned int);
     dtptr aggregate(DataTable*);
     static constexpr double epsilon = 1.0e-15;
-    static void set_norm_coeffs(double&, double&, double, double, int32_t);
+    static void set_norm_coeffs(double&, double&, double, double, size_t);
     static void print_progress(double, int);
 
   private:
-    int32_t min_rows;
-    int32_t n_bins;
-    int32_t nx_bins;
-    int32_t ny_bins;
-    int32_t nd_max_bins;
-    int32_t max_dimensions;
+    size_t min_rows;
+    size_t n_bins;
+    size_t nx_bins;
+    size_t ny_bins;
+    size_t nd_max_bins;
+    size_t max_dimensions;
     unsigned int seed;
     unsigned int nthreads;
     PyObject* progress_fn;
@@ -60,18 +60,18 @@ class Aggregator {
     void group_2d_mixed(bool, const dtptr&, dtptr&);
     template<typename T>
     void group_2d_mixed_str(bool, const dtptr&, dtptr&);
-    bool random_sampling(dtptr&, int32_t, int32_t);
+    bool random_sampling(dtptr&, size_t, size_t);
     void aggregate_exemplars(DataTable*, dtptr&, bool);
 
     // Helper methods
-    int32_t get_nthreads(const dtptr&);
+    size_t get_nthreads(const dtptr&);
     DoublePtr generate_pmatrix(const dtptr&);
-    void normalize_row(const dtptr&, DoublePtr&, int32_t);
-    void project_row(const dtptr&, DoublePtr&, int32_t, DoublePtr&);
-    double calculate_distance(DoublePtr&, DoublePtr&, int64_t, double,
+    void normalize_row(const dtptr&, DoublePtr&, size_t);
+    void project_row(const dtptr&, DoublePtr&, size_t, DoublePtr&);
+    double calculate_distance(DoublePtr&, DoublePtr&, size_t, double,
                               bool early_exit=true);
     void adjust_delta(double&, std::vector<ExPtr>&, std::vector<int64_t>&,
-                      int64_t);
+                      size_t);
     void adjust_members(std::vector<int64_t>&, dtptr&);
     size_t calculate_map(std::vector<int64_t>&, size_t);
     void progress(double, int status_code=0);
